2023.1 Release

+++ src/main/java/com/perforce/p4java/impl/generic/client/ClientOptions.java
@@ -17,6 +17,8 @@ public class ClientOptions implements IClientOptions {
 	private boolean locked = false;
 	private boolean modtime = false;
 	private boolean rmdir = false;
+	private boolean altsync = false;
+	private boolean altSyncFeatureEnabled = false;
 
 	/**
 	 * Default constructor; sets all fields to false.
@@ -44,8 +46,31 @@ public class ClientOptions implements IClientOptions {
 	}
 
 	/**
+	 * Explicit-value constructor.
+	 *
+	 * @param allWrite allWrite
+	 * @param clobber  clobber
+	 * @param compress compress
+	 * @param locked   locked
+	 * @param modtime  modtime
+	 * @param rmdir    rmdir
+	 * @param altsync  altsync
+	 */
+	public ClientOptions(boolean allWrite, boolean clobber, boolean compress, boolean locked, boolean modtime, boolean rmdir, boolean altsync) {
+		this.allWrite = allWrite;
+		this.clobber = clobber;
+		this.compress = compress;
+		this.locked = locked;
+		this.modtime = modtime;
+		this.rmdir = rmdir;
+		this.altsync = altsync;
+
+		this.altSyncFeatureEnabled = true;
+	}
+
+	/**
 	 * Attempts to construct a ClientOptions object from a typical p4 cmd options string,
-	 * e.g. "noallwrite noclobber nocompress unlocked nomodtime normdir". If optionsString
+	 * e.g. "noallwrite noclobber nocompress unlocked nomodtime normdir noaltsync". If optionsString
 	 * is null, this is equivalent to calling the default constructor.
 	 *
 	 * @param optionsString options
@@ -54,6 +79,9 @@ public class ClientOptions implements IClientOptions {
 
 		if (optionsString != null) {
 			String opts[] = optionsString.split(" ");
+
+			this.altSyncFeatureEnabled = isAltSyncFeatureEnabled(opts);
+
 			for (String str : opts) {
 				if (str.equalsIgnoreCase("allwrite")) {
 					this.allWrite = true;
@@ -67,9 +95,19 @@ public class ClientOptions implements IClientOptions {
 					this.modtime = true;
 				} else if (str.equalsIgnoreCase("rmdir")) {
 					this.rmdir = true;
+				} else if (str.equalsIgnoreCase("altsync")) {
+					this.altsync = true;
 				}
 			}
 		}
+	}
+
+	private boolean isAltSyncFeatureEnabled(String[] opts) {
+		for (String str : opts) {
+			if (str.equalsIgnoreCase("altsync") || str.equalsIgnoreCase("noaltsync"))
+				return true;
+		}
+		return false;
 	}
 
 	/**
@@ -78,7 +116,10 @@ public class ClientOptions implements IClientOptions {
 	 * constructor.
 	 */
 	public String toString() {
-		return (this.allWrite ? "allwrite" : "noallwrite") + (this.clobber ? " clobber" : " noclobber") + (this.compress ? " compress" : " nocompress") + (this.locked ? " locked" : " nolocked") + (this.modtime ? " modtime" : " nomodtime") + (this.rmdir ? " rmdir" : " normdir");
+		if (this.altSyncFeatureEnabled) //means [no]altsync is present and server version >= 2023.1
+			return (this.allWrite ? "allwrite" : "noallwrite") + (this.clobber ? " clobber" : " noclobber") + (this.compress ? " compress" : " nocompress") + (this.locked ? " locked" : " nolocked") + (this.modtime ? " modtime" : " nomodtime") + (this.rmdir ? " rmdir" : " normdir") + (this.altsync ? " altsync" : " noaltsync");
+		else
+			return (this.allWrite ? "allwrite" : "noallwrite") + (this.clobber ? " clobber" : " noclobber") + (this.compress ? " compress" : " nocompress") + (this.locked ? " locked" : " nolocked") + (this.modtime ? " modtime" : " nomodtime") + (this.rmdir ? " rmdir" : " normdir");
 	}
 
 	public boolean isAllWrite() {
@@ -127,5 +168,14 @@ public class ClientOptions implements IClientOptions {
 
 	public void setRmdir(boolean rmdir) {
 		this.rmdir = rmdir;
+	}
+
+	public boolean isaltSync() {
+		return altsync;
+	}
+
+	public void setAltsync(boolean altsync) {
+		this.altsync = altsync;
+		this.altSyncFeatureEnabled = true;
 	}
 }
